component rename correctly
